[JARenderer.java]: generateTilesVertices(): Changed algorithm in generateTilesVertices to use ymin + SIZE. Used the algorithm given to us in notes instead, which allows the squares to show up.
renderObjects():
Changed viewProjMatrix.setOrtho() from the values given for the previous assignment to the values given to us in the current assignment.
Changed VTD to be equal to indices.length.

test():
Added test function to view the indices and vertices. This was used to try and debug, since the squares weren't showing up properly.