[Driver.java]: Added driver for Assignment 8. This is just the driver from Assignment 7 copied over. No changes have been made yet.
[Test.java]:
Added test file to test whether the methods are working the way they are intended to work.

[JAPingPongArrayLive.java]:
Added file. Currently, constructs an object using the newly added empty constructor of JAPingPongArray so that I can access the variables and methods in that class. Additionally, it adds random live cells based on the liveCount passed through the constructor.

[JAPingPongArray.java]:
Added empty constructor.
Changed private variables and methods to be protected when needed, so I can access them from my subclass.
Changed the method getArray() to getLiveArray() and added an additional getNextArray().