[Assignment8_Test.java]: Added imports for java.io.* and java.util.*. I don't think I will need either of them, but it's just in case. Moved the JAPingPongArrayLive object creation to outside the main method so that I can use the public method calls with the object. Added code to ult_a(). Tests that loadFile() properly loads the file and that the array is being filled with the default value provided in the file. The test passes if all the values within both files match. Added code to ult_c(). This tests that the swapLiveAndNext() function works by making a string for the live array, swapping the arrays, making another string for the next array, and comparing the two arrays. The test passes if the two strings match.
[JAPingPongArray.java]:
Added getLiveString(). This returns the liveArray as a string for use in ult_c.
Added getNextString(). This returns the nextArray as a string for use in ult_c.
Added a line to loadFile() that takes the default value specified in the file and sets the object's defaultValue to that value.

[JAPingPongArrayLive.java]:
Added swapOnesAndZeros(). It changes the 1s and 0s in the nextArray() to 0 and 1 respectively.
Added resetNextArray(). This resets the nextArray to be full of 0s.
Added randomize(). It resets the nextArray and adds live values randomly into the array, then swaps the arrays and does it again. This is for use in ult_c to make sure that live and next array aren't the same, as that makes the results of ult_c inconclusive.

[Assignment8_Test.xml]:
File added when storing Assignment8_Test.java as a project file.