[Assignment8_Test.java]: Adjusted the nested for loops in ult_b() to use the rows and cols returned from myBoard2, rather than using the rows and cols defined at the start of the file. Looking back at my older assignment, I realized that while using rows and cols for the loops works, if I were to create a PingPongArray with different values for the rows and cols, the test would not function as expected.
[JAPingPongArray.java]:
Added comments to initializeArray() and getCols() to explain what the code is intended to do.
Altered code in getCols() to return COLS-1 rather than COLS because the first column is the row numbers which is not counted in the total columns (for example, if the object is created with 7 rows and 7 columns, the object will add a column for the row numbers).
Moved getNextArray() to be directly under getLiveArray(). I just thought it looked more organized having similar methods next to each other.

[JAPingPongArrayLive.java]:
Added a parameterless constructor. JAGoLArray is extending JAPingPongArrayLive, so it needs a parameterless constructor.
Corrected the comment for countLiveDegreeTwoNeighbors() where it incorrectly said that there would be 9 immediate squares around the initial 8 created with 1st degree, when the correct number is 12.

[Assignment9_test.java]:
Added file.

[Assignment9_test.xml]:
File created when storing Assignment9_test.java as a project file.

[Driver.java]:
Added file.

[JAGeometryManager.java]:
Added file. Created empty methods as per the professor's specs.

[JAGoLArray.java]:
Added file. Created empty methods as per the professor's specs.

[gol_input_1.txt]:
Added file. File will be used to test the program.