Assignment 9: [Assignment9_Test.java]: Added code to the test file. Currently, it creates two GoL boards, one using specific parameters and one using a text file. Then it calls testRun() which runs 10 iterations of the Game of Life to test the validity of the board updates.
[Driver.java]:
Added code. Currently creates a board and calls the run() function, which will run the game on an infinite loop once implemented.

[JAGoLArray.java]:
Changed liveCellCount to not have a default value and gave it a value when the object is created using the two parameter constructor. Currently assigns 50% of the board to be live when the user does not specify liveCellCount.
Added code to onTickUpdate(). When the method is called, it goes through the entire board one cell at a time and then changes the current cell to either be LIVE or DEAD depending on the number of live neighbors.
Added method updateCell(). This method takes the number of nearest neighbors as a parameter and returns LIVE or DEAD depending on the neighbor count. It returns LIVE if the cell has 3 live neighbors, and DEAD if the cell has either less than 2 or more than 3 neighbors.
Added method getLiveCount(). This method goes through the liveArray one cell at a time and returns the number of live cells. Originally, I simply deprecated liveCellCount when a cell was made dead. But it was unreliable because I did not increment liveCellCount when a cell was changed to be alive, which meant that liveCellCount could easily reach negative numbers. This current way is inefficient compared to using liveCellCount, so I might go back to using that to get the live cells count.
Added testRun() method. This method runs the same code as run(), but is specifically designed to show both the initial array before any updates happen and the live cell count each iteration. It only runs 11 iterations before stopping so that it is easier to tell what the program is doing every step.

[JAPingPongArrayLive.java]:
Changed all private methods to protected.
Added countLiveNeighbors() method. Since this assignment has us using only the 1st degree neighbors, this method was created to count the 1st degree neighbors.
Altered countLiveDegreeTwoNeighbors() method. Since I created the countLiveNeighbors() method, I removed the duplicated code and called countLiveNeighbors() within the method instead.
Added testRandomLive() method. This method is intended to test the addRandomLive() method and see whether or not it's adding the specified number of live cells to the array.

[JAPingPongArray.java]:
Altered printArray() to print out only the array itself without the row numbers.
Added printFullArray() to print out the full array, including the row numbers.
Altered the loadFile() method. I realized that the method did not work under certain conditions. If the user were able to load a file to construct the object, like in the case of this assignment, then neither the next nor live arrays are given a size and thus the program fails. I also realized that my previous code setting new variables rows and cols to the rows and columns specified in the file went unused and didn't do anything. Now, it sets ROWS and COLS to the values specified in the file and sets the live and next arrays to the new size.