Assignment 9: [JAGoLArray.java]: Separated the countLiveNeighbors() method call from the updateCell() method call to have a better idea of what countLiveNeighbors() is returning. Altered the parameters for updateCell(). Now takes the current status of the cell being checked. I was wondering why no matter what I did, before all 10 iterations of the test run, there would be no more live cells. I figured something had to be wrong, so while I was closely checking, I realized that since my updateCell() method starts with the retVal being DEAD, if there were exactly 2 neighbors, the cell would die. This goes against the rule where having two neighbors is supposed to retain the cell's status. So I fixed that with this change. Changed the testRun() to print the live count immediately after the "nth Iteration" line.
[JAPingPongArrayLive.java]:
Altered the conditionals within countLiveNeighbors() to require col to be greater than 1. I forgot to take the first column into consideration for this method. Since the first column is the row numbers, it should not be included. This change prevents the row number from having any effect on the 1st column in the board. I noticed this when cell [3,1] should have stayed dead, but came to life with only two nearby neighbors.
Similarly, altered the conditional within countLiveDegreeTwoNeighbors() to take this same issue into consideration.