Assignment 9:
[Assignment9_Test.java]:
Changed testRun() to test both the loadFile() and initial randomization of the board.
Added testRun2() to test the game using OpenGL. Most of the code was copied from Assignment 6's Driver.java.
Added spacing to ult methods.

[JAGeometryManager.java]:
Added variables to help with the creation of the squares and checking which squares should be rendered.
Changed JAPingPongArray to JAGoLArray. It was significantly harder trying to figure out how to make it work using the base PingPongArray functions. Since we need to check the live cells and I need the onTickUpdate() method for updating the board, I needed to change the reference to JAGoLArray.
Added code to the constructor. It just sets all the parameters to the values within the class.
Added code to generateTilesVertices(). I figured that this method is intended to call the generateTilesVertices(float[]), since otherwise we would have to consider multiple of the same class when calling it in the renderer object. I also removed the parameters, since I couldn't see how we are intended to use rowTiles and colTiles, considering that we are only supposed to render live tiles.
Added code to generateTilesVertices(float[]). This method does the majority of the work by filling in the actual vertices into the array. This code is mostly just copied over from the same method in the renderer class.
Added code to generateTilesIndices(). Like with the vertices method, I removed the parameter from this method as it wasn't being used. This code is mostly copied over from the same method in the renderer class.

[JAGoLArray.java]:
Removed loop from run(). The method was mostly to simplify the testing of the game in text format to make sure it worked. When moving over to the OpenGL version of the game, this loop and printing of the array is no longer needed.

[JAPingPongArray.java]:
Added getCell() method. This simplifies the process of checking the cell to ensure it's alive and giving it a square in the GeometryManager class.

[JARenderer.java]:
Added import for pkgJAUtils. Needed for a call to get the PingPongArray into the GeometryManager object.
Added GeometryManager object and SLEEP_INTERVAL variable.
Removed loop from renderObjects() method.
Changed vertices and indices method calls to be conditional based on the render() call we do. This allows Assignment 5 to still work.
Added a second render() method that takes an additional JAGoLArray object as a parameter.
Added code to renderLoop() to let the game run and update continuously at a specified tick rate.

Comments:
There were quite a few methods he gave us where I just could not figure out how I am intended to use them or their parameters. Take generateTilesVertices() for example. He gave us two versions of this method that pass two different parameter sets. But the initial one that passed two ints didn't make any sense to me, as we don't need to render every single tile. And fillArrayWithTileVertices doesn't make any sense either. My best guess is that we were supposed to use the (int,int) method to call the (float[]) method, which then calls the fillArrayWithTileVertices() method. But we aren't really splitting responsibilities between these three methods.

I'm almost, but not quite, done with this assignment. I think overall it's a good assignment, but too much confusion is left for the students to deal with, especially considering that he just gives us the UML, the parameter signatures, and that's it. If I had an idea of what fillArrayWithTileVertices() was supposed to do, maybe I would have used that method. But as it stands, I didn't use it, and had no reason to use it.