Assignment 9:
[Assignment9_Test.java]:
Added testRun3(). This test was to figure out why the OpenGL version would sometimes have cells that seemed to break the rules and stay alive near the right edge of the screen. This test gives us the raw values of the array, which allowed me to see that the cells were following the rules at the edge of the screen and that it was a rendering issue, rather than a game logic issue.
Added ult_e(). This test is to ensure that the rules of the game are being followed by checking a file after a single automated update with a file that is the first array after being updated manually.

[Driver.java]:
Added code to the driver. Now runs the program. This code is mostly copied over from testRun2() from the test file.

[JAGeometryManager.java]:
Changed the lines in the for loops of both generateTilesVertices() and generateTilesIndices(). Now creates vertices and indices up to NUM_COLS rather than NUM_COLS-1.

[JAGoLArray.java]:
Changed the file constructor. Now initializes the liveArray after loading the file.
Changed the two JAGoLArray(int,int) constructor. Now calls the super(int,int,int) constructor with the parameters, rather than simply calling super() and setting the parameters.

[JARenderer.java]:
Changed SLEEP_INTERVAL from 40 to 50.

[ppa_data.txt]:
Added file. File was added through the save() method and through the results of my testing.

[ult_e_input.txt]:
Added file. File was added to test ult_e.

[ult_e_verify.txt]:
Added file. File was added to test ult_e.