Assignment 9:
[Assignment9_Test.java]:
Changed runTests() to take a String as a parameter to be used in ult_h.
Added three more unit level tests.
Added ult_f to test the creation and destruction of an OpenGL window.
Added ult_g to test the save(String) method.
Added initial bits to ult_h, which will take a string as a parameter and create a board using it. The string will come from the command line arguments.

[Assignment9_Test.xml]:
Changed the configuration to include a command line argument.

[ult_g_verify.txt]:
File added through ult_g, which tests creation of a file using the save(String) method.