Assignment 9:
[Assignment9_Test.java]:
Changed the main method to pass the argument, if there is one. If an argument is given, it's passed to the runTests() function. Else, null is passed.
Added ult_h. This tests that the command line argument matches the string "gol_input_1.txt".
Added bits of ult_i. ult_h was initially intended to test the creation of a board using the command line argument, but I thought it would be better to first test that the command line argument was being passed correctly first. Now, ult_i will test the creation of the board with the command line argument.