Assignment 9:
[Assignment9_Test.java]:
Changed runTests() to notify the user that they are missing the command line argument needed for ult_h and ult_i.
Changed ult_h to return out of the test if s == null. Since s.equals() will return a runtime error if s == null, we need to account for that and allow the program to continue running if no string is passed into the command line.
Changed ult_i to return out of the test of s == null. This is for the same reason as ult_h.
Added ult_j to test the creation of a board without setting a live count. This is to test that the board will set its own live count value equal to 20% of the cells if the user does not provide one.

[JAGeometryManager.java]:
Added comments to remind myself how the indices should be passed. One of the problems I had with this class was it not properly passing the indices.

[JAGoLArray.java]:
Added comments to further explain what some of the methods do.

[JAPingPongArray.java]:
Added findCSC133() method. This method calls searchForCSC133() to try to find the correct folder which houses the various text files used with loadFile() and save(). One of the issues my programs had when being graded by the professor or the graders was that the program was searching in the wrong folder for the files. This happened because the root folder was being changed when the grader extracted my program files.
Added searchForCSC133() method. This method searches through the directory and subdirectories until it finds the CSC133 folder.
Altered loadFile() to create the filepath according to the result of findCSC133().
Moved loadFile() to be closer to the save() method. This was both for organization and making it easier to bugfix, as this bug affected both loadFile() and save().
Altered save() to create the filepath according to the result of findCSC133().

[JAPingPongArrayLive.java]:
Added comments to further explain some of the methods and what some of the code is doing.
Removed comments that were unnecessary, such as commented out code.

[JARenderer.java]:
Removed the commented out code.
Added a comment to explain the conditional in the generate vertices and indices section of renderObjects().