Assignment 9:
[JAGeometryManager.java]:
Added comments to explain some of the methods.

[JAGoLArray.java]:
Added further comments to liveCount().

[JAPingPongArray.java]:
Added comments to explain some of the methods and the algorithms used in them.

[JAPingPongArrayLive.java]:
Added comments to explain the countLiveNeighbors() method.

[gol_input_3.txt]:
Added file. I liked the patterns this array produced, so I saved it as a file.